Adding Stream in separate file
