+++ sdk/core/azure-core/inc/http/http.hpp
@@ -3,6 +3,8 @@
 
 #pragma once
 
+#include "stream.h"
+
 #include <algorithm>
 #include <internal/contract.hpp>
 #include <map>
@@ -14,33 +16,9 @@ namespace Azure { namespace Core { namespace Http {
 
   enum class TransportKind
   {
-      //TODO move this to Factory
-      Curl,
-      WinHttp
-  };
-
-  // BodyStream is used to read data to/from a service
-  class BodyStream {
-  public:
-    static BodyStream* null;
-
-    // Returns the length of the data; used with the HTTP Content-Length header
-    virtual uint64_t Length() = 0;
-
-    // Resets the stream back to the beginning (for retries)
-    // Derived classes that send data in an HTTP request MUST override this and implement it
-    // properly.
-    virtual void Rewind()
-    {
-      throw "Not Implemented"; // TODO: Replace with best practice as defined by guideline
-    };
-
-    // Reads more data; EOF if return < count; throws if error/canceled
-    virtual uint64_t Read(/*Context& context, */ uint8_t* buffer, uint64_t offset, uint64_t count)
-        = 0;
-
-    // Closes the stream; typically called after all data read or if an error occurs.
-    virtual void Close() = 0;
+    // TODO move this to Factory
+    Curl,
+    WinHttp
   };
 
   enum class HttpStatusCode
